UI Designs

+++ Lab10/Lab10/db4odemo/src/userinterface/PatientRole/OrderDetailsJPanel.java
@@ -51,7 +51,7 @@ public class OrderDetailsJPanel extends javax.swing.JPanel {
     {
     
         DefaultTableModel model = (DefaultTableModel) OrderTable.getModel();
-        int Price=0;
+        float Price=0;
         model.setRowCount(0);
         for (Order order : o) {
